requirements.txt
numpy
scikit-learn
python-dotenv
streamlit
faiss-cpu
google-generativeai
PyPDF2
sentence-transformers
transformers
